pyproject.toml
[project]
name = "deployer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "uvicorn>=0.35.0",
]
